pyproject.toml
[project]
name = "rag-with-langchain"
version = "0.1.0"
description = ""
authors = [
    {name = "manuel orlandi",email = "manu.orlandi91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "trafilatura >=2.0.0",
    "polars >=1.26.0",
    "langchain (>=0.3,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "instructorembedding (>=1.0.1,<2.0.0)",
    "sentence-transformers (>=4.0.1,<5.0.0)",
    "streamlit (>=1.44.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]